# [qa] blockstore: Switch to dumb dbm