Closes #8605 

I could not see any performance degradation by this commit:


```sh
$ time qa/pull-tester/rpc-tests.py p2p-fullblocktest
```

Before:
```
TEST                 | PASSED | DURATION

p2p-fullblocktest.py | True   | 297 s

ALL                  | True   | 297 s (accumulated)

Runtime: 297 s

real	4m57.633s
user	4m1.588s
sys	0m11.419s
```

After:

```
p2p-fullblocktest.py:
Pass: True, Duration: 288 s

TEST                 | PASSED | DURATION

p2p-fullblocktest.py | True   | 288 s

ALL                  | True   | 288 s (accumulated)

Runtime: 288 s

real	4m48.573s
user	3m56.126s
sys	0m11.823s