@laanwj 2016-09-29
Concept ACK, this is by far the most straightforward solution, if there is no need for platform-specific exceptions then let's avoid them.