@laanwj 2016-09-29
Interesting, apparently this causes the NULLDUMMY test to fail during close, on some platforms. I don't understand why this file would suddenly be missing during close:
```
stderr:
Exception ignored in: <bound method _Database.close of <dbm.dumb._Database object at 0x7f8f097f1be0>>
Traceback (most recent call last):
  File "/usr/lib/python3.4/dbm/dumb.py", line 250, in close
    self._commit()
  File "/usr/lib/python3.4/dbm/dumb.py", line 112, in _commit
    with self._io.open(self._dirfile, 'w', encoding="Latin-1") as f:
FileNotFoundError: [Errno 2] No such file or directory: '/tmp/testkz7n1z0q/1/blocks.dir'
Exception ignored in: <bound method _Database.close of <dbm.dumb._Database object at 0x7f8f097f1d68>>
Traceback (most recent call last):
  File "/usr/lib/python3.4/dbm/dumb.py", line 250, in close
    self._commit()
  File "/usr/lib/python3.4/dbm/dumb.py", line 112, in _commit
    with self._io.open(self._dirfile, 'w', encoding="Latin-1") as f:
FileNotFoundError: [Errno 2] No such file or directory: '/tmp/testkz7n1z0q/1/transactions.dir'
```